Final:
AN-2001
Owner
AN-2001 commented now
Rework how transformations are handled to make it more in line with irit grap drivers.
Remove installation system and add scripts.
Add logger.
Handle lights better.
Add ZMin/ZMax.
Add zOffset.
Add screenshot system and write offscreen rendrer.

+++ Irit2Powerpoint/WinAPIDclr.cs
@@ -46,13 +46,13 @@ namespace Irit2Powerpoint
 
         public delegate IntPtr WinProc(IntPtr hWnd, uint msg, IntPtr wParam, IntPtr lParam); 
 
-        public static int LOWORD(IntPtr Num)
+        public static short LOWORD(IntPtr Num)
         {
-            return (int)Num & 0xffff;
+            return (short)((int)Num & 0xffff);
         }
-        public static int HiWORD(IntPtr Num)
+        public static short HiWORD(IntPtr Num)
         {
-            return ((int)Num >> 16) & 0xffff;
+            return (short)(((int)Num >> 16) & 0xffff);
         }
     }
 }
